+++ test.projects/test/java/test/projects/JigsawQuickStartWorldTests.java
@@ -15,11 +15,11 @@ class JigsawQuickStartWorldTests {
     var out =
         cli.start(
             Command.of("bach")
-                .add("--verbose")
-                .add("--strict")
-                .add("--limit-tools", "javac,jar")
-                .add("--jar-with-sources")
-                .add("build"));
+                .with("--verbose")
+                .with("--strict")
+                .with("--limit-tools", "javac,jar")
+                .with("--jar-with-sources")
+                .with("build"));
     assertLinesMatch(
         """
         >> BACH'S INITIALIZATION >>
@@ -45,7 +45,7 @@ class JigsawQuickStartWorldTests {
         """
             .lines()
             .sorted(),
-        CLI.run(Command.jar().add("--list").add("--file", greetings)).lines().sorted());
+        CLI.run(Command.jar().with("--list").with("--file", greetings)).lines().sorted());
     var world = cli.workspace("modules", "org.astro@0.jar");
     assertLinesMatch(
         """
@@ -60,6 +60,6 @@ class JigsawQuickStartWorldTests {
         """
             .lines()
             .sorted(),
-        CLI.run(Command.jar().add("--list").add("--file", world)).lines().sorted());
+        CLI.run(Command.jar().with("--list").with("--file", world)).lines().sorted());
   }
 }