+++ test.projects/test/java/test/projects/MultiRelease9Tests.java
@@ -17,11 +17,11 @@ class MultiRelease9Tests {
     var out =
         cli.start(
             Command.of("bach")
-                .add("--verbose")
-                .add("--project-targets-java", "9")
-                .add("--limit-tools", "javac,jar")
-                .add("--jar-with-sources")
-                .add("build"));
+                .with("--verbose")
+                .with("--project-targets-java", "9")
+                .with("--limit-tools", "javac,jar")
+                .with("--jar-with-sources")
+                .with("build"));
     assertLinesMatch(
         """
         >> BACH'S INITIALIZATION >>
@@ -60,6 +60,6 @@ class MultiRelease9Tests {
         """
             .lines()
             .sorted(),
-        CLI.run(Command.jar().add("--list").add("--file", foo)).lines().sorted());
+        CLI.run(Command.jar().with("--list").with("--file", foo)).lines().sorted());
   }
 }